requirements.txt
# Policy-to-Code Pipeline Requirements

# Core dependencies (required)
# No required dependencies for basic functionality!

# Azure OpenAI Integration (optional)
openai>=1.0.0

# Environment variable management (optional but recommended)
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
